refactor: Phase 4 - Move chat, grammar, transliteration features to ui/ packages
- Move all chat components (ChatActivity, ChatViewModel, ChatAdapter, etc.) to ui/chat/
- Move grammar components (GrammarActivity, GrammarViewModel, GrammarCorrectionAdapter) to ui/grammar/
- Move TransliterationActivity to ui/transliteration/
- Move SubscriptionUIComponent to ui/subscription/
- Update all package declarations and imports
