+++ app/src/main/java/com/kharagedition/tibetankeyboard/ui/chat/ChatAdapter.kt
@@ -1,4 +1,4 @@
-package com.kharagedition.tibetankeyboard.chat
+package com.kharagedition.tibetankeyboard.ui.chat
 import android.content.ClipData
 import android.content.Context
 import android.view.LayoutInflater
