+++ app/src/main/java/com/kharagedition/tibetankeyboard/ui/chat/ChatHistoryAdapter.kt
@@ -1,4 +1,4 @@
-package com.kharagedition.tibetankeyboard.chat
+package com.kharagedition.tibetankeyboard.ui.chat
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
