+++ app/src/main/java/com/kharagedition/tibetankeyboard/ui/chat/ChatViewModel.kt
@@ -1,4 +1,4 @@
-package com.kharagedition.tibetankeyboard.chat
+package com.kharagedition.tibetankeyboard.ui.chat
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
